// packages/core/src/native/registry/Registry.types.ts
export type Uint128 = string
export interface InstantiateMsg {
  admin: string
  allow_direct_module_registration_and_updates?: boolean | null
  namespace_registration_fee?: Coin | null
}
export interface Coin {
  amount: Uint128
  denom: string
}
// ...
export type Namespace = string
export type ModuleVersion =
  | 'latest'
  | {
      version: string
    }
export type ModuleReference =
  | {
      account_base: number
    }
  | {
      native: Addr
    }
  | {
      adapter: Addr
    }
  | {
      app: number
    }
  | {
      standalone: number
    }
export type Addr = string
export type UpdateModule =
  | {
      default: {
        metadata: string
      }
    }
  | {
      versioned: {
        instantiation_funds?: Coin[] | null
        metadata?: string | null
        monetization?: Monetization | null
        version: string
      }
    }
export type Monetization =
  | 'none'
  | {
      install_fee: FixedFee
    }
// ...
export interface ModuleInfo {
  name: string
  namespace: Namespace
  version: ModuleVersion
}
export interface FixedFee {
  fee: Coin
}
// ...
export interface ModuleResponse {
  config: ModuleConfiguration
  module: Module
}
export interface ModuleConfiguration {
  instantiation_funds: Coin[]
  metadata?: string | null
  monetization: Monetization
}
export interface Module {
  info: ModuleInfo
  reference: ModuleReference
}